Fix modalizer examples  add additional logging
This PR fixes the modalizer examples by initializing the Modalizer and
Deloser APIs before the stories are rendered. This was not caught
previously during the migration from `demo.tsx` beceause those APIs were
initialized on the global scope.

Also adds some additional documentation and console.error development
output to warn about this in the future