Remove modalizer accessibility and set aria-hidden in tree instead (#27)
* Unique modalizer instance per user id

This PR refactors modalizer so that one single instance exists and
manages and Modalizer DOM elements per user id. This makes sense since
all modalizers with the same id need to behave the same way.

This will enable more logic to be moved from `Root` to `Modalizer` in
the future. Lots of logic currently in `Root` regarding modalizer exists
because of the limitation that multiple modalizer instances can exist
that have the same user id.

* add comment

* simplify root

* update root

* remove nodefilter

* Experiment: Remove modalizer accessibility

This PR refactors modalizers to set `aria-hidden` on all nodes that do
not belong to the specific modalizer.

Removes the concept of accessibility of modalizers and cleans up Root
API to be unaware of modalizers

* comment modalizer usage in deloser

* cleanup more modalizers

* lint and comments

* remove modalizer timer

* improve story

* improve story

* modalizer obeserve mutations

* perf optimization

* remove modalizer ref from Focusable

* add comments

* uncomment code

* chage modalizer to single element and stop using main modalizers

* add modalizer conditions to accept element

* remove useless import

* leverage curModalizer

* lint

* extract fuction

* add comments

* enhance stories

* enhance stories

* enhance stories

* fix restore focus bug

* rename variable

* rename fix

* Remove active modalizer on mutation

* improve comments

* add isOthersAccessible example

* simplify tree walking

* don't use imperative add APIs