Merging 0.7.0 changes to master (#32)
* Remove modalizer accessibility and set aria-hidden in tree instead (#27)

* Unique modalizer instance per user id

This PR refactors modalizer so that one single instance exists and
manages and Modalizer DOM elements per user id. This makes sense since
all modalizers with the same id need to behave the same way.

This will enable more logic to be moved from `Root` to `Modalizer` in
the future. Lots of logic currently in `Root` regarding modalizer exists
because of the limitation that multiple modalizer instances can exist
that have the same user id.

* add comment

* simplify root

* update root

* remove nodefilter

* Experiment: Remove modalizer accessibility

This PR refactors modalizers to set `aria-hidden` on all nodes that do
not belong to the specific modalizer.

Removes the concept of accessibility of modalizers and cleans up Root
API to be unaware of modalizers

* comment modalizer usage in deloser

* cleanup more modalizers

* lint and comments

* remove modalizer timer

* improve story

* improve story

* modalizer obeserve mutations

* perf optimization

* remove modalizer ref from Focusable

* add comments

* uncomment code

* chage modalizer to single element and stop using main modalizers

* add modalizer conditions to accept element

* remove useless import

* leverage curModalizer

* lint

* extract fuction

* add comments

* enhance stories

* enhance stories

* enhance stories

* fix restore focus bug

* rename variable

* rename fix

* Remove active modalizer on mutation

* improve comments

* add isOthersAccessible example

* simplify tree walking

* don't use imperative add APIs

* A test.

* Oops.

* Keyborg package. (#28)

* v0.7.0-alpha.0

* chore(keyborg): Cleaup and dedupe (#29)

* chore(keyborg): Cleaup and dedupe

Adds docstrings and renames some things for readability

Dedupes multipe `WeakRef` classes by creating a weak ref generic wrapper

* revert underscore

* remove unnecessary comments

* enhance comment

* Fix(Modalizer): Dispose inactive modalizers (#30)

* Fix(Modalizer): Dispose inactive modalizers

Currently the mutation listener for `Modalizer` only disposes inactive
modalizers. This is a problem for 'stacks' of `Modalizers` (i.e. when
one modalizer opens another one)

When all modalizers in a stack are removed from DOM, some may already
inactive so doesn't get cleaned up.

This changes makes sure that any `Modalizer` removed from DOM will be
dsposed from the `ModalizerAPI`

* fix story

* improve onFocus

* Fix(autoRoot): auto root should have tabster attr (#31)

* Fix(autoRoot): auto root should have tabster attr

Currently `autoRoot` does not have tabster attributes applied to the DOM
element. This results in mutations not being dispatched to the `RootAPI`
because `getTabsterOnElement` is used on the mutation target before the
observer decides to dispatch the mutation event

* remove tabster on element with root is disposed

* Root instance sets/removes tabster on elemt

* v0.7.0-alpha.1

* fix(demo): Replace awesome-typescript-loader with ts-loader (#33)

Fixes the error in s-panferov/awesome-typescript-loader #372

* fix(Focusable): Groupper condition in _acceptElement was inverted

Fixes a bug introduced in #27 when the `_acceptElement` function was
modified

Co-authored-by: Marat Abdullin <marata@microsoft.com>