fix(Modalizer+Deloser): Check if already focused (#38)
* fix(Modalizer): Check if a new modalizer is already focused

Currently `ModalizerAPI.add` does not check if the modalizer is already
focused. Therefore the only way to activate a modalizer is through
`Modalizer.API.focus()` or listening to the focus event in Tabster.

This leads to a bug if a new modalizer is added that already has focus.
This can be caused by React's virtual DOM rendering which can't
guarantee the order of mutation event and focus event. Here is an
example:

* add tabster modalizer attributes to DOM element
* focus DOM element
* `ModalizerAPI._onFocus` triggers first and can't find a modalizer on the element ->
does nothing
* `Instance` adds the modalizer
* Focus is on the modalizer, but it is not active and not other content
on the page is hidden

* activate new deloser

* remove hasFocus in favour of tabster focusedElement