fix(keyboardNavigationState): use `nativeFocus` when simulating keyboard
The `nativeFocus` is the original focus before keybord hijacks it, but
also serves a special use case in the library.

Using `nativeFocus` will allow keyborg to detect that we are trying to
simulate keyboard behaviour, which will change the value of the keyboard
navigation state.

This unknown aspect was ignored in #64 and regressed, where Mover and
Groupper APIs were not updating keyboard navigation state correctly