Moves keyborg code into `/core` and removes keyborg package
As discussed offline, all uses of keyborg are in a context where tabster
is also used. Moves keyborg package back into tabster. This can be
republished at a later date when it is required.

The problem `keyborg` was solving initially was to reduce the impact of
using Tabster's keyboard navigation state on the bundle size. Since
treee shaking is supported, the new package is not necessary