fix: Dispose `__tabsterDummy` on DummyInputManagerCore
When a `DummyInputManagerCore` is created it attaches itself to a HTML
element. When it is disposed it does not remove itself from the HTML
element.

When a new tabster core is created it will still use the undisposed
`DummyInputManagerCore` and causes a leak where a disposed tabster core
is being used