feat(sys): System API for adjusting the dummy inputs position in rare cases of necessity. (#239)
In uncontrolled mode we redirect focus using dummy inputs (invisible focusable elements whose only purpose is to move focus to the proper element once focused). These dummy inputs observe their position in DOM and adjust it when the application changes its DOM. Depending on the use case the dummy inputs can be positioned inside the container element or outside of it.

For example, if we have Mover on a <div>, the dummy inputs will be created as first and last element in the <div>, but if we have Mover on a <table>, the dummy inputs will be created outside (because having them inside is invalid HTML and causes glitches).

Generally the placement of the dummy inputs is not the application creator's concern. However, there are corner cases when we want to change the exact placement. One of the examples is when we have Mover on a <div> and the application keeps updating the DOM inside that <div>, Tabster will reposition the dummy inputs after every DOM update (to make sure that the dummy inputs are first and last in the <div>). Depending on application structure and logic, it might be a performance concern (because changing the DOM causes the reflow) and putting the dummy inputs outside of the <div> might be a way to mitigate the problem.

We are adding sys API for the internal Tabster's logic adjustments like this. It's not a part of Mover, Groupper and Modalizer because it doesn't belong there semantically and because of possible conflicting declarations when, for example, Groupper and Modalizer are used on the same element.

Another change is the adjustment of how the data-tabster properties are passed to the constructors internally (which doesn't affect public API, but makes the pull request a bit bigger). Now the entire set of props is available inside the constructors, not just the relevant to the constructor part.