feat(DummyInput): Using only main mutation observer to update dummy inputs in DOM and bulk-updating transform offsets (#254)
Two optimizations regarding dummy inputs:

There is no mutation observed in DummyInputManagerCore anymore. We're utilizing the main mutation observer to keep the dummy inputs DOM placement consistent. And the DOM placement update is throttled to reduce the amount of reflows.

Every DummyInputManagerCore doesn't update the dummy input visual position on scroll independently, it is being queued to do the bulk-update in order to reduce the amount of reflows.