fix(DummyInput): Force dummy input position update when it is needed before the scheduled update. (#278)
Dummy inputs are supposed to be first/last elements in the container. When new DOM is appended after the dummy input, Tabster will schedule a position update on the next tick. In the real world it is hardly possible to press Tab between the DOM update and the scheduled dummy input update, but in the test environment it is easily possible. Because the position affects the behaviour like moveOutWithDefaultAction(), we enforce its update from the move out functions.