fix: Restorer should add target to history on creation if focused
In the cases where the actual restorer is not created yet and the target
is focused - it will not be in history. This is because tabster does not
exist on the element yet.

Handle this case specifically on creation and add a new target element
to history if it is already the active element