fix(Mover): Fixing the case when focusable mover container was causing memorization of current element on the wrong mover. (#325)
In case of nested movers, when the inner mover's container is focusable, the current element was memorized on that inner mover, not on the parent mover. Because of that memorizeCurrent behaviour of the parent mover didn't work properly.

When the movers are nested and something in the inner mover receives focus, the parent movers were not populated with the proper current element.