feat(Uncontrolled): completely prop to indicate that uncontrolled area fully controls Tab. (#327)
Currently we have checkUncontrolledTrappingFocus callback in createTabster() to provide the flag that an uncontrolled area is trapping focus. Adding completely property to uncontrolled props on DOM element to be able to set this indicator right on the particular DOM elements. Also deprecating checkUncontrolledTrappingFocus and adding checkUncontrolledCompletely for the sake of consistency and better coexistence.