feat(Deloser): Preventable events to control Deloser, manual focus restoratin Deloser strategy. (#361)
Deloser now dispatches TabsterMoveFocusEvent before restoring the focus. It is possible for application to preventDefault() this event.

Adding strategy prop to Deloser with new Manual value. Manual delosers won't restore focus automatically, but they still follow the focus history.

It is possible for the application to dispatch DeloserRestoreFocusEvent to force the Deloser to restore focus.