feat(ObservedElement): Adding observed element request status. (#425)
When requestFocus() request resolves with true or false, in case of false it was not possible to tell the difference between timed out request or canceled request (older requests get canceled when a new request is made or when a user has moved focus). Adding status to be able to tell one from another.