fix(Modalizer): Do not follow virtual parents when removing aria-hidden for the element focused in modalizer. (#460)
When active modalizer is changed, we augment aria-hidden attributes asynchronously for the performance reasons. But we synchronously remove currently augmented aria-hidden attributes from the focused element parents (so that screen reader is able to see it right away). This synchronous aria-hidden unaugmentation should not follow virtual parents, just the real ones.