fix(Mover): Explicitly adding combobox role to elements that should be ignored by Mover when aria-expanded=true. (#486)
Mover tries to be smart in detecting if the focus should be moved with the arrow key press or not. One of the cases when it should not be moved is when there is an expanded menu (like combobox). We were checking it by the presence of the pair of aria-expanded and aria-activedescendant attributes. Combobox might allow the cases when it is expanded, but nothing is chosen via aria-activedescendant because the combobox items do not match the value that is currently typed in the combobox input. So, we are extending the check to explicitly check for a combobox role.